Move ui application loader in service repository
