+++ ui/main/src/app/services/applicationLoader/ApplicationLoader.ts
@@ -8,43 +8,43 @@
  */
 
 import {firstValueFrom} from 'rxjs';
-import {ConfigService} from '../services/config/ConfigService';
+import {ConfigService} from '../config/ConfigService';
 import {LoggerService as logger, LogOption} from 'app/services/logs/LoggerService';
-import {RemoteLoggerService} from '../services/logs/RemoteLoggerService';
-import {TranslationService} from '../services/translation/TranslationService';
-import {UsersService} from '../services/users/UsersService';
-import {OpfabAPI} from '../api/opfab.api';
-import {GlobalStyleService} from '../services/style/global-style.service';
-import {EntitiesService} from '../services/entities/EntitiesService';
-import {GroupsService} from '../services/groups/GroupsService';
-import {PerimetersService} from '../services/perimeters/PerimetersService';
-import {ProcessesService} from '../services/processes/ProcessesService';
-import {OpfabEventStreamService} from '../services/events/OpfabEventStreamService';
-import {AcknowledgeService} from '../services/acknowlegment/AcknowledgeService';
-import {DateTimeFormatterService} from '../services/dateTimeFormatter/DateTimeFormatterService';
-import {AdminProcessesService} from '../services/admin/AdminProcessesService';
-import {BusinessDataService} from '../services/businessdata/businessdata.service';
-import {CardsService} from '../services/cards/CardsService';
-import {SupervisedEntitiesService} from '../services/admin/SupervisedEntitiesService';
-import {HandlebarsService} from '../services/handlebars/HandlebarsService';
-import {ExternalDevicesService} from '../services/notifications/ExternalDevicesService';
-import {TemplateCssService} from '../services/templateCss/TemplateCssService';
-import {UserSettingsService} from '../services/userSettings/UserSettingsService';
-import {NavigationService} from '../services/navigation/NavigationService';
-import {OpfabStore} from '../store/opfabStore';
-import {ApplicationUpdateService} from '../services/events/ApplicationUpdateService';
-import {SystemNotificationService} from '../services/notifications/SystemNotificationService';
-import {ApplicationLoadingComponent} from './application-loading-component';
-import {ServerResponseStatus} from './server/serverResponse';
-import {Utilities} from '../utils/utilities';
-import {ModalService} from '../services/modal/ModalService';
-import {SessionManagerService} from '../services/sessionManager/SessionManagerService';
-import {SoundNotificationService} from '../services/notifications/SoundNotificationService';
+import {RemoteLoggerService} from '../logs/RemoteLoggerService';
+import {TranslationService} from '../translation/TranslationService';
+import {UsersService} from '../users/UsersService';
+import {OpfabAPI} from '../../api/opfab.api';
+import {GlobalStyleService} from '../style/global-style.service';
+import {EntitiesService} from '../entities/EntitiesService';
+import {GroupsService} from '../groups/GroupsService';
+import {PerimetersService} from '../perimeters/PerimetersService';
+import {ProcessesService} from '../processes/ProcessesService';
+import {OpfabEventStreamService} from '../events/OpfabEventStreamService';
+import {AcknowledgeService} from '../acknowlegment/AcknowledgeService';
+import {DateTimeFormatterService} from '../dateTimeFormatter/DateTimeFormatterService';
+import {AdminProcessesService} from '../admin/AdminProcessesService';
+import {BusinessDataService} from '../businessdata/businessdata.service';
+import {CardsService} from '../cards/CardsService';
+import {SupervisedEntitiesService} from '../admin/SupervisedEntitiesService';
+import {HandlebarsService} from '../handlebars/HandlebarsService';
+import {ExternalDevicesService} from '../notifications/ExternalDevicesService';
+import {TemplateCssService} from '../templateCss/TemplateCssService';
+import {UserSettingsService} from '../userSettings/UserSettingsService';
+import {NavigationService} from '../navigation/NavigationService';
+import {OpfabStore} from '../../store/opfabStore';
+import {ApplicationUpdateService} from '../events/ApplicationUpdateService';
+import {SystemNotificationService} from '../notifications/SystemNotificationService';
+import {ApplicationLoadingComponent} from './ApplicationLoadingComponent';
+import {ServerResponseStatus} from '../../business/server/serverResponse';
+import {Utilities} from '../../utils/utilities';
+import {ModalService} from '../modal/ModalService';
+import {SessionManagerService} from '../sessionManager/SessionManagerService';
+import {SoundNotificationService} from '../notifications/SoundNotificationService';
 import {I18n} from '@ofModel/i18n.model';
-import {RealTimeDomainService} from '../services/realTimeDomain/RealTimeDomainService';
-import {NotificationDecision} from '../services/notifications/NotificationDecision';
-import {CardTemplateGateway} from '../services/templateGateway/CardTemplateGateway';
-import {UserCardTemplateGateway} from '../services/templateGateway/UserCardTemplateGateway';
+import {RealTimeDomainService} from '../realTimeDomain/RealTimeDomainService';
+import {NotificationDecision} from '../notifications/NotificationDecision';
+import {CardTemplateGateway} from '../templateGateway/CardTemplateGateway';
+import {UserCardTemplateGateway} from '../templateGateway/UserCardTemplateGateway';
 import {SelectedCardService} from '@ofServices/selectedCard/SelectedCardService';
 import {UserActionLogsService} from '@ofServices/userActionLogs/UserActionLogsService';
